#91 move Page component
