+++ src/pages/Landing/LandingPage.tsx
@@ -2,7 +2,7 @@ import { Navigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 
 import { useAuth } from 'common/hooks/useAuth';
-import Page from 'common/components/Page/Page';
+import Page from 'common/components/Content/Page';
 
 /**
  * The `LandingPage` component renders the content of the landing page
